refactor: break views functionality into more specific modules; cleanup generated code; remove unused files
