+++ packages/plugin-views/src/components/FieldEnhancer/index.tsx
@@ -5,8 +5,11 @@ import type { Field } from "@puckeditor/core";
 
 import getClassNameFactory from "../../../../core/lib/get-class-name-factory";
 import { useCurrentNodeEditor } from "../../hooks/use-current-node-editor";
-import { getNodeViewState, setNodeViewState } from "../../lib/views";
-import type { NodeViewState, ViewsPluginOptions } from "../../types";
+import {
+  getNodeViewState,
+  setNodeViewState,
+} from "../../lib/bindings";
+import type { ViewsPluginOptions } from "../../types";
 
 import { BindingControl } from "../BindingControl";
