+++ packages/plugin-views/src/components/ObjectField/index.tsx
@@ -1,6 +1,6 @@
 import { memo, ReactNode, useMemo } from "react";
-import { AutoField, FieldLabel, ObjectField as PuckObjectField } from "@/core";
-import { getClassNameFactory } from "@/core/lib";
+import { AutoField, FieldLabel, ObjectField as PuckObjectField } from "@puckeditor/core";
+import getClassNameFactory from "@/core/lib/get-class-name-factory";
 
 import styles from "./style.module.css";
