+++ packages/plugin-views/src/hooks/use-puck.ts
@@ -1,4 +1,4 @@
-import { createUsePuck } from "@/core";
+import { createUsePuck } from "@puckeditor/core";
 
 /**
  * Hook to access the Puck internal API using selectors to minimize re-renders.
