+++ packages/plugin-views/src/lib/constants.ts
@@ -1,0 +1,8 @@
+/** Default storage key for views in root props */
+export const DEFAULT_STORAGE_KEY = "__puck_views";
+
+/** Default storage key for node state in component props */
+export const DEFAULT_NODE_STATE_KEY = "__puck_view_state";
+
+/** Default key in metadata to signal if the plugin should add data binding options or not */
+export const RENDER_DATA_BINDING_KEY = "__puck_add_data_binding";
